feat(rounds_payload): change UpdateCompetitionConfig handler to use Rounds Batch Patch
move the handling of CompetitionRounds from PutCompetitionRounds to
BatchPatchCompetitionRounds

We generate two paths for handling the eventId.

Path one handles the fake-event id key for brand new events and includes
that in the update which translates to a PUT. This is necessary to have
the eventId present on creation because of the Dynamo GSI

The second path omits the eventId of events that have an ID present than
the placeholder. This prevents the update from overwriting the eventId
from the competitionConfig page. It prevents edits there from getting
the attribute out of sync.