feat(rounds_payload): move payload from Add Selected Rounds to Publish Event
This adds formData.pendingRoundUpdates field to store rounds. It removes
the direct call to update competitionRounds when the "Add Selected
Rounds" button is pressed.

the handleAddRoundClick now only stores the rounds locally in state and
on the formData so that it is passed in the handlePublishClick payload
in order to only add the eventIds to the rounds