feat(monolisth_dockerization): dockerfile and supervisord config file
This introduces a docker container that runs a pre-compiled go binary
with the necessary dependencies in a streamlined alpine image. It
includes supervisord to run the Go process and lay ground work for
running future processes such as weaviate and postgres. It copies
necessary env file into the working directory in order to load env vars.