feat(monolith_dockerization_phase1): move to docker compose
This commit moves toward docker compose for running the postgres,
weaviate and the go application in the separate containers. It includes
a specific golang app Dockerfile and currently runs standard weaviate
and postgres17 images. This also includes updates to have supervisord
only run in our golang container to assist retreats for hot reloading
the go app. Also, included is the goWatch.js script that is modelled
after the watchTempl.js script.