feat(gh_package_workflow): change to Dockerfile.go.prod for GH packages
Using the Dockerfile.go.prod for the Building and Published docker
image. The suffixed dev dockerfile is not setup for deploying to a
server but only local development. The .prod Dockerfile has the build
stage and streamlined final container and should have no issues or
differences between a prod instance and a dev instance.