feat(weaviate_seeding): a script to seed weaviate from json
This introduces the test event dataset (it is messy) but as pulled from
the dev server.

It also introduces the cmd seed-weaviate-db that can be run with the
--file=<json_file>  flag in order to seed the db. This can be used for
dev or prod.

There are the two commands found in the package.json for create schema
and upserting documents.

Updates to README that are in progress and a couple little fixes in the
dev dockerfile for local development